moved samplespec and added files property for importing raw files
