+++ src/Components/SampleEditor/SampleEditor.tsx
@@ -1,0 +1,4 @@
+import { File } from "@bentley/monaco-editor";
+export interface InternalFile extends File {
+  import: Promise<{ default: string; }>;
+}
