+++ src/frontend-samples/heatmap-decorator-sample/index.tsx
@@ -3,7 +3,6 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import * as React from "react";
-import { SampleSpec } from "../../Components/SampleShowcase/SampleShowcase";
 import { GithubLink } from "../../Components/GithubLink";
 import "../../common/samples-common.scss";
 import { IModelApp, Viewport, StandardViewId, IModelConnection } from "@bentley/imodeljs-frontend";
@@ -13,22 +12,9 @@ import { Range2d, Point3d } from "@bentley/geometry-core";
 import HeatmapDecorator from "./HeatmapDecorator";
 import { ColorDef } from "@bentley/imodeljs-common";
 import { PointSelector } from "../../common/PointSelector/PointSelector";
-import { SampleIModels } from "../../Components/IModelSelector/IModelSelector";
-
-export function getHeatmapDecoratorSpec(): SampleSpec {
-  return ({
-    name: "heatmap-decorator-sample",
-    label: "Heatmap Decorator",
-    image: "heatmap-decorator-thumbnail.png",
-    modelList: [SampleIModels.RetailBuilding, SampleIModels.BayTown, SampleIModels.House],
-    handlesViewSetup: true,
-    setup: HeatmapDecoratorApp.setup,
-    teardown: HeatmapDecoratorApp.teardown,
-  });
-}
 
 /** This class implements the interaction between the sample and the iModel.js API.  No user interface. */
-class HeatmapDecoratorApp {
+export class HeatmapDecoratorApp {
 
   public static decorator?: HeatmapDecorator;
   public static range?: Range2d;
