+++ src/frontend-samples/viewport-only-sample/index.tsx
@@ -4,23 +4,11 @@
 *--------------------------------------------------------------------------------------------*/
 import * as React from "react";
 import "@bentley/icons-generic-webfont/dist/bentley-icons-generic-webfont.css";
-import { SampleSpec } from "../../Components/SampleShowcase/SampleShowcase";
 import { GithubLink } from "../../Components/GithubLink";
 import "../../common/samples-common.scss";
 import { IModelConnection, Viewport } from "@bentley/imodeljs-frontend";
-import { SampleIModels } from "../../Components/IModelSelector/IModelSelector";
 
 // cSpell:ignore imodels
-
-export function getViewportOnlySpec(): SampleSpec {
-  return ({
-    name: "viewport-only-sample",
-    label: "Viewport Only",
-    image: "viewport-only-thumbnail.png",
-    modelList: [SampleIModels.RetailBuilding, SampleIModels.BayTown, SampleIModels.House],
-    setup: async (imodel: IModelConnection, vp: Viewport) => { return <ViewportOnlyUI iModel={imodel} vp={vp} /> },
-  });
-}
 
 export class ViewportOnlyUI extends React.Component<{ iModel: IModelConnection, vp: Viewport }, {}> {
