+++ src/raw-loader.d.ts
@@ -1,0 +1,4 @@
+declare module '!!raw-loader!*' {
+  const contents: string
+  export = contents
+}
